Update spring tests

+++ backend/rest_api/SentimentAnalysis/src/test/java/com/sentimentanalysis/SentimentAnalysis/SentimentAnalysisApplicationTests.java
@@ -54,8 +54,8 @@ public class SentimentAnalysisApplicationTests {
 		this.mockMvc.perform(get("/api/search?query=java")).andDo(print())
 				.andExpect(status().isOk())
 				.andExpect(jsonPath("$.rating", is(notNullValue())))
-				.andExpect(jsonPath("$.rating", greaterThanOrEqualTo(-1.0)))
-				.andExpect(jsonPath("$.rating", lessThanOrEqualTo(1.0)))
+				.andExpect(jsonPath("$.rating", greaterThanOrEqualTo(BigDecimal.valueOf(-1.0))))
+				.andExpect(jsonPath("$.rating", lessThanOrEqualTo(BigDecimal.valueOf(1.0))))
 				.andExpect(jsonPath("$.rating", is(notNullValue())))
 				.andExpect(jsonPath("$.posts").isArray())
 				.andExpect(jsonPath("$.posts[0].*", hasSize(4)))
@@ -63,8 +63,8 @@ public class SentimentAnalysisApplicationTests {
 				.andExpect(jsonPath("$.posts[0].timestamp", not(isEmptyOrNullString())))
 				.andExpect(jsonPath("$.posts[0].url", not(isEmptyOrNullString())))
 				.andExpect(jsonPath("$.posts[0].score", is(notNullValue())))
-				.andExpect(jsonPath("$.posts[0].score", greaterThanOrEqualTo(-1.0)))
-				.andExpect(jsonPath("$.posts[0].score", lessThanOrEqualTo(1.0)));
+				.andExpect(jsonPath("$.posts[0].score", greaterThanOrEqualTo(Double.valueOf(-1.0))))
+				.andExpect(jsonPath("$.posts[0].score", lessThanOrEqualTo(Double.valueOf(1.0))));
 	}
 
 	@Test
@@ -74,8 +74,8 @@ public class SentimentAnalysisApplicationTests {
 				.andExpect(jsonPath("$.timestamps").isArray())
 				.andExpect(jsonPath("$.timestamps[0]", not(isEmptyOrNullString())))
 				.andExpect(jsonPath("$.medians").isArray())
-				.andExpect(jsonPath("$.medians[0]", greaterThanOrEqualTo(-1.0)))
-				.andExpect(jsonPath("$.medians[0]", lessThanOrEqualTo(1.0)));
+				.andExpect(jsonPath("$.medians[0]", greaterThanOrEqualTo(Double.valueOf(-1.0))))
+				.andExpect(jsonPath("$.medians[0]", lessThanOrEqualTo(Double.valueOf(1.0))));
 	}
 
 }
